pyproject.toml
[project]
name = "bunq-hack-2027"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "streamlit>=1.35.0",
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.0",
  "python-multipart>=0.0.9",
  "boto3>=1.34.0",
  "httpx[http2]>=0.27.0",
  "python-dotenv>=1.0.0",
  "requests>=2.31.0",
  "cryptography>=41.0.0",
  "aws-sdk-bedrock-runtime>=0.5.0",
  "smithy-aws-core>=0.5.0",
  # fraud detection
  "anthropic>=0.30.0",
  "opencv-python-headless>=4.9.0",
  "numpy>=1.26.0",
  "Pillow>=10.0.0",
]
